refactor(bench): extract shared write-cost measure and Db-factory helpers
Two copy-paste duplications in bench/ tooling collapsed into shared
modules (bench-only code, no shipped surface):

- CF_FREE/NODE maintenance options, the Workload interface, and the
  measure() harness are now in bench/lib/write-cost-measure.ts; the
  amortized and stress write-cost scripts keep only their WORKLOADS
  arrays and reporting. The bootstrap owner label is a measure()
  parameter so each script keeps its own tag, and the stress reporter
  rebuilds its row object in the original key order to keep its
  baseline JSON byte-identical.
- The per-tenant Db cache (dbFor) shared byte-for-byte by the seed and
  replay runners is now makeDbFactory() in
  bench/load-harness/runner/db-factory.ts.

Pure refactor; no behavior change. Baselines reproduce byte-for-byte.
